+++ ImprovedStorageInfo/ImprovedStorageInfo/Patches/SeamothStorageInputPatches.cs
@@ -1,6 +1,6 @@
 ﻿using HarmonyLib;
 
-namespace Koi.Subnautica.ImprovedStorageInfo.core
+namespace Koi.Subnautica.ImprovedStorageInfo.Patches
 {
     /// <summary>
     /// The root harmony patched for Seamoth Storage Input game object.
@@ -17,13 +17,13 @@ namespace Koi.Subnautica.ImprovedStorageInfo.core
         // ReSharper disable once InconsistentNaming
         public static void OnHandHover(SeamothStorageInput __instance)
         {
-            if (!ModPlugin.ConfigEnabled.Value) return;
+            if (!ModConfig.ConfigEnabled.Value) return;
 
             var itemContainer = __instance.seamoth.GetStorageInSlot(__instance.slotID, TechType.VehicleStorageModule);
 
             HandReticle.main.SetText(
                 HandReticle.TextType.HandSubscript,
-                ContainerUtils.GetCustomInteractText(itemContainer),
+                Utils.ContainerUtils.GetCustomInteractText(itemContainer),
                 false
             );
         }