+++ ImprovedScanInfo/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-### [1.0.0] 2023-03-15
+## [1.1.0] 2023-03-19
+
+### Functionnalities
+
+* Improve translation system by reload translations on game language changes
+
+### Technical
+
+* Remove external assembly for logger
+
+## [1.0.0] 2023-03-15
 
 * Initial commit