+++ ImprovedStorageInfo/CHANGELOG.md
@@ -3,18 +3,28 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-### [1.0.0] 2023-03-14
+## [2.1.0] 2023-03-19
 
-* Initial commit
+### Functionnalities
 
-### [2.0.0] 2023-03-15
+* Improve translation system by reload translations on game language changes
 
-#### Technical
+### Technical
+
+* Remove external assembly for logger
+
+## [2.0.0] 2023-03-15
+
+### Technical
 
 * Use external assembly for logger
 * Use external assembly for translations
 
-#### Fixes
+### Fixes
 
 * Missing display for escape pod container
 * Missing display for vehicule containers
+
+## [1.0.0] 2023-03-14
+
+* Initial commit